Moving overlay to PB context
